update pareto example
